pages/requirements.txt
PyPDF2==3.0.1
beautifulsoup4==4.12.3
scikit-learn==1.5.2
openai
streamlit==1.39.0
python-dotenv==1.0.1
langchain==0.3.2
langchain-openai==0.2.2
langchain-google-genai==2.0.1
langchain-perplexity==0.1.0
langchain-community==0.3.0
pylatex==1.4.2
pillow
pytesseract
reportlab

replicate
pdf2image
svgwrite
python-docx 
Pillow
pdf2docx
markdown